+++ lib/src/utils/supported_locales_utils.dart
@@ -17,14 +17,12 @@ class SupportedLocalesUtils {
     'cs': CsLocale(),
     'da': DaLocale(),
     'de': DeLocale(),
-    'dv': DvLocale(),
     'en': EnLocale(),
     'es': EsLocale(),
     'et': EtLocale(),
     'fa': FaLocale(),
     'fi': FiLocale(),
     'fr': FrLocale(),
-    'gr': GrLocale(),
     'he': HeLocale(),
     'hi': HiLocale(),
     'hr': HrLocale(),
@@ -35,7 +33,6 @@ class SupportedLocalesUtils {
     'ka': KaLocale(),
     'km': KmLocale(),
     'ko': KoLocale(),
-    'ku': KuLocale(),
     'lv': LvLocale(),
     'mn': MnLocale(),
     'mn_MY': MsMyLocale(),
@@ -43,18 +40,15 @@ class SupportedLocalesUtils {
     'my': MyLocale(),
     'nb': NbLocale(),
     'nl': NlLocale(),
-    'nn': NnLocale(),
     'pl': PlLocale(),
     'ps': PsLocale(),
     'pt': PtLocale(),
     'ro': RoLocale(),
     'ru': RuLocale(),
-    'rw': RwLocale(),
     'sr': SrLocale(),
     'sv': SvLocale(),
     'ta': TaLocale(),
     'th': ThLocale(),
-    'tk': TkLocale(),
     'tl_PH': TlPhLocale(),
     'tr': TrLocale(),
     'uk': UkLocale(),
@@ -64,14 +58,21 @@ class SupportedLocalesUtils {
     'zh_CN': ZhCnLocale(),
   };
 
+  static final Map<String, Locale> _onlySupportedRelatives = {
+    'dv': DvLocale(),
+    'gr': GrLocale(),
+    'ku': KuLocale(),
+    'nn': NnLocale(),
+    'rw': RwLocale(),
+    'tk': TkLocale(),
+  };
+
+  static final Map<String, Locale> _mergedSupportedLocales = {
+    ..._supportedLocales,
+    ..._onlySupportedRelatives,
+  };
+
   /// Returns a [Locale] instance for the given locale code.
-  /// If the given locale is not supported, it will return the default locale.
-  /// If the given locale is null, it will return the default locale.
-  /// If the given locale is empty, it will return the default locale.
-  /// If the given locale is not found, it will return the default locale.
-  /// If the given locale is found, it will return the locale.
-  /// If the given locale is found, but the country code is not found, it will
-  /// return the locale without the country code.
   static Locale getLocale(String? locale) {
     final defaultLocale = locale ?? Intl.defaultLocale;
 
@@ -85,6 +86,21 @@ class SupportedLocalesUtils {
     return defaultLocaleSupport;
   }
 
+  /// Returns a [Locale] instance for the given locale code, including those
+  /// that only support relative date formatting.
+  static Locale getRelativeLocale(String? locale) {
+    final defaultLocale = locale ?? Intl.defaultLocale;
+
+    final defaultLocaleSupport = _supportedLocales['en']!;
+    if (defaultLocale != null && defaultLocale.isNotEmpty) {
+      final finalLocale = defaultLocale.split('_').first;
+      return _mergedSupportedLocales[locale] ??
+          _mergedSupportedLocales[finalLocale] ??
+          defaultLocaleSupport;
+    }
+    return defaultLocaleSupport;
+  }
+
   /// Returns `true` if the given locale is supported.
   static bool isLocaleSupported(String locale) {
     return _supportedLocales.containsKey(locale.toLowerCase());
@@ -93,5 +109,10 @@ class SupportedLocalesUtils {
   /// Returns a list of supported locales.
   static List<String> getSupportedLocales() {
     return _supportedLocales.keys.toList();
+  }
+
+  /// Returns a list of supported relative locales.
+  static List<String> getSupportedRelativeLocales() {
+    return _mergedSupportedLocales.keys.toList();
   }
 }